use namespace in config generator
