+++ lib/generators/spid_rails/templates/spid_rails.rb
